+++ packages/web-vue/components/upload/upload-list-item.tsx
@@ -68,7 +68,7 @@ export default defineComponent({
         <div class={`${itemCls}-content`}>
           {uploadCtx?.listType === 'picture' && (
             <span class={`${itemCls}-thumbnail`}>
-              <img src={props.file.url} />
+              <img src={props.file.url} alt={props.file.name} loading={uploadCtx.imageLoading} />
             </span>
           )}
           <div class={`${itemCls}-name`}>
@@ -80,7 +80,7 @@ export default defineComponent({
             )}
             {uploadCtx?.customIcon?.fileName?.(props.file) ||
               (props.file.url ? (
-                <a class={`${itemCls}-name-link`} href={props.file.url}>
+                <a class={`${itemCls}-name-link`} target="_blank" href={props.file.url} download={props.file.name}>
                   {props.file.name}
                 </a>
               ) : (
@@ -101,11 +101,11 @@ export default defineComponent({
           <UploadProgress file={props.file} listType={props.listType} />
         </div>
         <span class={`${itemCls}-operation`}>
-          <IconHover onClick={() => uploadCtx?.onRemove?.(props.file)}>
+          {uploadCtx?.showRemoveButton && <IconHover onClick={() => uploadCtx?.onRemove?.(props.file)}>
             <span class={[uploadCtx?.iconCls, `${uploadCtx?.iconCls}-remove`]}>
               {uploadCtx?.customIcon?.removeIcon?.() || <IconDelete />}
             </span>
-          </IconHover>
+          </IconHover>}
         </span>
       </div>
     );