+++ packages/web-vue/components/upload/upload-list.tsx
@@ -4,6 +4,7 @@ import UploadListItem from './upload-list-item';
 import UploadPictureItem from './upload-picture-item';
 import { FileItem, ListType } from './interfaces';
 import { isFunction } from '../_utils/is';
+import { uploadInjectionKey } from './context';
 
 export default defineComponent({
   name: 'UploadList',
@@ -20,6 +21,10 @@ export default defineComponent({
       type: String as PropType<ListType>,
       required: true,
     },
+    isMax: {
+      type: Boolean,
+      default: false,
+    },
   },
   setup(props, { slots }) {
     const prefixCls = getPrefixCls('upload');
@@ -28,6 +33,8 @@ export default defineComponent({
       `${prefixCls}-list`,
       `${prefixCls}-list-type-${props.listType}`,
     ]);
+
+    const uploadCtx = inject(uploadInjectionKey, undefined);
 
     const renderItem = (fileItem: FileItem, index: number) => {
       if (isFunction(slots['upload-item'])) {
@@ -49,7 +56,7 @@ export default defineComponent({
     return () => (
       <TransitionGroup tag="div" class={cls.value}>
         {props.fileList.map((item, index) => renderItem(item, index))}
-        {props.listType === 'picture-card' && slots['upload-button']?.()}
+        {props.listType === 'picture-card' && !props.isMax && uploadCtx?.showUploadButton && slots['upload-button']?.()}
       </TransitionGroup>
     );
   },