+++ packages/web-vue/components/upload/upload.tsx
@@ -158,6 +158,50 @@ export default defineComponent({
       default: true,
     },
     /**
+     * @zh 是否显示删除按钮
+     * @en Whether to display the remove button
+     * @version 2.11.0
+     */
+    showRemoveButton: {
+      type: Boolean,
+      default: true,
+    },
+    /**
+     * @zh 是否显示重试按钮
+     * @en Whether to display the retry button
+     * @version 2.11.0
+     */
+    showRetryButton: {
+      type: Boolean,
+      default: true,
+    },
+    /**
+     * @zh 是否显示取消按钮
+     * @en Whether to display the cancel button
+     * @version 2.11.0
+     */
+    showCancelButtoon: {
+      type: Boolean,
+      default: true,
+    },
+    /**
+     * @zh 是否显示上传按钮
+     * @en Whether to display the retry button
+     * @version 2.11.0
+     */
+    showUploadButton: {
+      type: Boolean,
+      default: true,
+    },
+    /**
+     * @zh <img> 的原生属性
+     * @en <img> native attribute
+     * @version 2.11.0
+     */
+    imageLoading: {
+      type: String as PropType<'eager' | 'lazy'>,
+    },
+    /**
      * @zh 图片列表类型
      * @en Picture list type
      * @values 'text','picture','picture-card'
@@ -171,7 +215,9 @@ export default defineComponent({
      * @en Get the key of the image URL in the Response. After opening, it will replace the pre-load image with the uploaded image
      */
     responseUrlKey: {
-      type: String,
+      type: [String, Function] as PropType<
+        string | ((fileItem: FileItem) => string)
+      >,
     },
     /**
      * @zh 自定义图标
@@ -349,8 +395,12 @@ export default defineComponent({
           file.status = 'done';
           file.percent = 1;
           file.response = response;
-          if (props.responseUrlKey && response[props.responseUrlKey]) {
-            file.url = response[props.responseUrlKey];
+          if (props.responseUrlKey) {
+            if (isFunction(props.responseUrlKey)) {
+              file.url = props.responseUrlKey(file)
+            } else if (response[props.responseUrlKey]) {
+              file.url = response[props.responseUrlKey];
+            }
           }
 
           requestMap.delete(file.uid);
@@ -508,6 +558,10 @@ export default defineComponent({
         disabled,
         listType,
         iconCls: `${prefixCls}-icon`,
+        showRemoveButtoon: props.showRemoveButtoon,
+        showRetryButton: props.showRetryButton,
+        showCancelButtoon: props.showCancelButtoon,
+        imageLoading: props.imageLoading,
         customIcon,
         onUpload: uploadFile,
         onAbort: abort,
@@ -538,7 +592,7 @@ export default defineComponent({
 
     const render = () => {
       if (!props.showFileList) {
-        return renderButton();
+        return props.showUploadButton && renderButton();
       }
 
       return (
@@ -548,7 +602,7 @@ export default defineComponent({
             `${prefixCls}-wrapper-type-${props.listType}`,
           ]}
         >
-          {props.listType !== 'picture-card' && renderButton()}
+          {props.listType !== 'picture-card' && props.showUploadButton && renderButton()}
           <UploadList
             v-slots={{
               'upload-button': renderButton,
@@ -556,6 +610,8 @@ export default defineComponent({
             }}
             fileList={_fileList.value}
             listType={props.listType}
+            isMax={isMax.value}
+            showUploadButton={props.showUploadButton}
           />
         </div>
       );